Update README.md
C programming language is a general-purpose, procedural, high-level programming language used to develop computer software, computer applications, system programming, game development, web development, and more. C language was developed by Dennis M. Ritchie at the Bell Telephone Laboratories in 1972. C is a powerful and flexible language. It is a widely used programming language. C was developed for the programming of the UNIX operating System.

C programming is known for its simplicity and efficiency. It is a better choice to start with programming, It gives you a foundational understanding of programming.